refactor(usb-drive): extract platform types and partition helpers
Prevents cycles and moves the partition helpers to the level they belong.
