+++ libs/usb-drive/src/index.ts
@@ -9,4 +9,5 @@ export * from './types';
 export * from './usb_drive';
 export * from './usb_drive_adapter';
 export * from './usb_platform';
+export * from './usb_platform_types';
 export { type MockFileTree, writeMockFileTree } from './mocks/helpers';
